requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pyairtable>=2.3.0
google-auth>=2.35.0
google-api-python-client>=2.150.0
trafilatura>=1.12.0
pydantic>=2.9.0
python-dotenv>=1.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
beautifulsoup4>=4.12.0
